refactor: reorganize location module internal package structure
Restructure the flat internal package into organized sub-packages for
better separation of concerns and maintainability:

- `controller/` - REST controllers and exception handler
- `service/` - business logic and exceptions
- `repository/` - data access layer
- `entities/` - JPA entity classes
- `dtos/request/` - incoming request DTOs
- `dtos/response/` - outgoing response DTOs and mappers

Split `LocationEvents.kt` into individual event files and `LocationDtos.kt`
into separate per-class files. Fix `nLocation` class name typo to `Location`.
