+++ src/main/kotlin/com/nickdferrara/fitify/location/LocationDeactivatedEvent.kt
@@ -1,0 +1,9 @@
+package com.nickdferrara.fitify.location
+
+import java.time.Instant
+import java.util.UUID
+
+data class LocationDeactivatedEvent(
+    val locationId: UUID,
+    val effectiveDate: Instant,
+)
