+++ src/main/kotlin/com/nickdferrara/fitify/location/LocationUpdatedEvent.kt
@@ -1,0 +1,8 @@
+package com.nickdferrara.fitify.location
+
+import java.util.UUID
+
+data class LocationUpdatedEvent(
+    val locationId: UUID,
+    val updatedFields: Set<String>,
+)
