+++ src/main/kotlin/com/nickdferrara/fitify/location/internal/controller/LocationController.kt
@@ -1,5 +1,7 @@
-package com.nickdferrara.fitify.location.internal
+package com.nickdferrara.fitify.location.internal.controller
 
+import com.nickdferrara.fitify.location.internal.dtos.response.LocationResponse
+import com.nickdferrara.fitify.location.internal.service.LocationService
 import org.springframework.http.ResponseEntity
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
