+++ src/main/kotlin/com/nickdferrara/fitify/location/internal/dtos/response/ErrorResponse.kt
@@ -1,0 +1,3 @@
+package com.nickdferrara.fitify.location.internal.dtos.response
+
+internal data class ErrorResponse(val message: String)
