+++ src/main/kotlin/com/nickdferrara/fitify/location/internal/dtos/response/OperatingHoursResponse.kt
@@ -1,0 +1,20 @@
+package com.nickdferrara.fitify.location.internal.dtos.response
+
+import com.nickdferrara.fitify.location.internal.entities.LocationOperatingHours
+import java.time.DayOfWeek
+import java.time.LocalTime
+import java.util.UUID
+
+internal data class OperatingHoursResponse(
+    val id: UUID,
+    val dayOfWeek: DayOfWeek,
+    val openTime: LocalTime,
+    val closeTime: LocalTime,
+)
+
+internal fun LocationOperatingHours.toResponse() = OperatingHoursResponse(
+    id = id!!,
+    dayOfWeek = dayOfWeek,
+    openTime = openTime,
+    closeTime = closeTime,
+)
