+++ src/main/kotlin/com/nickdferrara/fitify/location/internal/entities/Location.kt
@@ -1,4 +1,4 @@
-package com.nickdferrara.fitify.location.internal
+package com.nickdferrara.fitify.location.internal.entities
 
 import jakarta.persistence.CascadeType
 import jakarta.persistence.Column
