+++ src/main/kotlin/com/nickdferrara/fitify/location/internal/service/LocationService.kt
@@ -1,10 +1,17 @@
-package com.nickdferrara.fitify.location.internal
+package com.nickdferrara.fitify.location.internal.service
 
 import com.nickdferrara.fitify.location.LocationApi
 import com.nickdferrara.fitify.location.LocationCreatedEvent
 import com.nickdferrara.fitify.location.LocationDeactivatedEvent
 import com.nickdferrara.fitify.location.LocationSummary
 import com.nickdferrara.fitify.location.LocationUpdatedEvent
+import com.nickdferrara.fitify.location.internal.dtos.request.CreateLocationRequest
+import com.nickdferrara.fitify.location.internal.dtos.request.UpdateLocationRequest
+import com.nickdferrara.fitify.location.internal.dtos.response.LocationResponse
+import com.nickdferrara.fitify.location.internal.dtos.response.toResponse
+import com.nickdferrara.fitify.location.internal.entities.Location
+import com.nickdferrara.fitify.location.internal.entities.LocationOperatingHours
+import com.nickdferrara.fitify.location.internal.repository.LocationRepository
 import com.nickdferrara.fitify.shared.DomainError
 import com.nickdferrara.fitify.shared.NotFoundError
 import com.nickdferrara.fitify.shared.Result
