+++ src/test/kotlin/com/nickdferrara/fitify/location/internal/LocationServiceTest.kt
@@ -3,6 +3,14 @@ package com.nickdferrara.fitify.location.internal
 import com.nickdferrara.fitify.location.LocationCreatedEvent
 import com.nickdferrara.fitify.location.LocationDeactivatedEvent
 import com.nickdferrara.fitify.location.LocationUpdatedEvent
+import com.nickdferrara.fitify.location.internal.dtos.request.CreateLocationRequest
+import com.nickdferrara.fitify.location.internal.dtos.request.OperatingHoursRequest
+import com.nickdferrara.fitify.location.internal.dtos.request.UpdateLocationRequest
+import com.nickdferrara.fitify.location.internal.entities.Location
+import com.nickdferrara.fitify.location.internal.entities.LocationOperatingHours
+import com.nickdferrara.fitify.location.internal.repository.LocationRepository
+import com.nickdferrara.fitify.location.internal.service.LocationNotFoundException
+import com.nickdferrara.fitify.location.internal.service.LocationService
 import com.nickdferrara.fitify.shared.NotFoundError
 import com.nickdferrara.fitify.shared.Result
 import io.mockk.every
